requirements.txt
langchain-ollama
langchain-core
langchainhub
langchain-community
langchain-google-genai
python-dotenv
